apps/scoring/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
Pillow==10.3.0
qrcode==7.4.2
